[WIP] Fast standalone symbolize for unwinding
We've had issues using addr2line. On certain versions of
CentOS it is on a version that has a performance regression making it very slow,
and even normallly it is not that fast, taking several seconds even when parallelized
for a typical memory trace dump.

LLVMSymbolize is fast but it requires PyTorch take an LLVM dependency
which we have been trying to remove.

This adds a standalone symbolizer to PyTorch similar to the unwinder which has
not external dependencies and is ~20x faster than addr2line for unwinding PyTorch frames.

ghstack-source-id: de078a4ea51b08a0d732a02e88b6c8cebbea822d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123966